Added drag and drop uploading
Added drag and drop image uploading.
Updated styles to show proper errors.
Made the uploaders JSON endpoints instead of pages.
Need to add an alternative for non ajax enabled clients.
Or make it all an AJAX API and turn the site in a SPA?
I think the upload successful message is pretty slick! I should make it
a plugin haha
Maybe add that animated character for the drag leave event? That would
be pretty funny
Still need to add deleting and anonymous tagging.